Completed order and payment function Start to do: 1. Check order in member page 2. Fix minor thing

+++ gulimall-order/src/main/java/com/_yzhheng/rest/dto/OmsOrderDTO.java
@@ -11,7 +11,6 @@ import java.math.BigDecimal;
 
 import java.util.Date;
 
-
 /**
  * DTO class for "OmsOrder"
  *
@@ -21,59 +20,57 @@ import java.util.Date;
 @Data
 public class OmsOrderDTO implements Serializable {
 
-    private static final long serialVersionUID = 1L;
+  private static final long serialVersionUID = 1L;
 
-    //--- PRIMARY KEY 
-    private Long id ;
-    //--- OTHER DATA FIELDS 
-    private Long memberId ;
-    private String orderSn ;
-    private Long couponId ;
-    private Date createTime ;
-    private String memberUsername ;
-    private BigDecimal totalAmount ;
-    private BigDecimal payAmount ;
-    private BigDecimal freightAmount ;
-    private BigDecimal promotionAmount ;
-    private BigDecimal integrationAmount ;
-    private BigDecimal couponAmount ;
-    private BigDecimal discountAmount ;
-    private Byte payType ;
-    private Byte sourceType ;
-    private Byte status ;
-    private String deliveryCompany ;
-    private String deliverySn ;
-    private Integer autoConfirmDay ;
-    private Integer integration ;
-    private Integer growth ;
-    private Byte billType ;
-    private String billHeader ;
-    private String billContent ;
-    private String billReceiverPhone ;
-    private String billReceiverEmail ;
-    private String receiverName ;
-    private String receiverPhone ;
-    private String receiverPostCode ;
-    private String receiverProvince ;
-    private String receiverCity ;
-    private String receiverRegion ;
-    private String receiverDetailAddress ;
-    private String note ;
-    private Byte confirmStatus ;
-    private Byte deleteStatus ;
-    private Integer useIntegration ;
-    private Date paymentTime ;
-    private Date deliveryTime ;
-    private Date receiveTime ;
-    private Date commentTime ;
-    private Date modifyTime ;
+  // --- PRIMARY KEY
+  private Long id;
+  // --- OTHER DATA FIELDS
+  private Long memberId;
+  private String orderSn;
+  private Long couponId;
+  private Date createTime;
+  private String memberUsername;
+  private BigDecimal totalAmount;
+  private BigDecimal payAmount;
+  private BigDecimal freightAmount;
+  private BigDecimal promotionAmount;
+  private BigDecimal integrationAmount;
+  private BigDecimal couponAmount;
+  private BigDecimal discountAmount;
+  private Byte payType;
+  private Byte sourceType;
+  private Integer status;
+  private String deliveryCompany;
+  private String deliverySn;
+  private Integer autoConfirmDay;
+  private Integer integration;
+  private Integer growth;
+  private Byte billType;
+  private String billHeader;
+  private String billContent;
+  private String billReceiverPhone;
+  private String billReceiverEmail;
+  private String receiverName;
+  private String receiverPhone;
+  private String receiverPostCode;
+  private String receiverProvince;
+  private String receiverCity;
+  private String receiverRegion;
+  private String receiverDetailAddress;
+  private String note;
+  private Integer confirmStatus;
+  private Integer deleteStatus;
+  private Integer useIntegration;
+  private Date paymentTime;
+  private Date deliveryTime;
+  private Date receiveTime;
+  private Date commentTime;
+  private Date modifyTime;
 
-    /**
-     * Constructor
-     */
-    public OmsOrderDTO() {
-		super();
-    }
+  /**
+   * Constructor
+   */
+  public OmsOrderDTO() {
+    super();
+  }
 }
-    
-
